LayoutTests/inspector/timeline-mark-timeline-expected.txt

Tests the Timeline API mark feature

MarkTimeline Properties:
+ startTime : * DEFINED *
+ data : {
+- message : MARK TIMELINE
+ }
+ type : 11
